make reducer from story part call
