+++ botstory/ast/story_context/__init__.py
@@ -1,5 +1,7 @@
 from botstory import matchers
-from botstory.ast import callable, forking, stack_utils
+from botstory.ast import callable, forking
+from botstory.ast.story_context import reducers
+
 import json
 import logging
 
@@ -70,57 +72,10 @@ class StoryContext:
     def to_json(self):
         return {
             'message': self.message,
-            'waiting_for': self.waiting_for,
+            'waiting_for':  str(self.waiting_for),
         }
 
     def __repr__(self):
         return json.dumps(self.to_json())
 
-
-# Context Reducers
-
-# TODO: should make it immutable
-def scope_in(ctx):
-    """
-    - build new scope on the top of stack
-    - and current scope will wait for it result
-
-    :param ctx:
-    :return:
-    """
-    compiled_story = None
-    if not ctx.is_empty_stack():
-        compiled_story = ctx.get_child_story()
-        # TODO: ! use reduceer
-        last_stack_item = ctx.stack_tail()
-        last_stack_item['step'] += 1
-        last_stack_item['data'] = matchers.serialize(callable.WaitForReturn())
-
-    if not compiled_story:
-        compiled_story = ctx.compiled_story()
-
-    logger.debug('# [>] going deeper')
-    # TODO: ! use reduceer
-    stack = ctx.stack()
-    stack.append(stack_utils.build_empty_stack_item(compiled_story.topic))
-
-    return ctx
-
-
-# TODO: should make it immutable
-def scope_out(ctx):
-    """
-    drop last stack item if we have reach the end of stack
-    and don't wait any input
-
-    :param ctx:
-    :return:
-    """
-    # we reach the end of story line
-    # so we could collapse previous scope and related stack item
-    if ctx.is_tail_of_story() and not ctx.is_waiting_for_input():
-        logger.debug('# [<] return')
-        # TODO: !
-        ctx.stack().pop()
-
-    return ctx
+__all__ = [reducers, StoryContext]
